refactor: extract RocketChatPresenceClient from RocketChatService
